@@ -1,10 +1,12 @@
-module.exports = {
-  future: {},
-  purge: [],
-  darkMode: false,
-  theme: {
-    extend: {},
-  },
-  variants: {},
-  plugins: [],
+/** @type {import('tailwindcss').Config} */
+export default {
+	content: [
+		"./index.html",
+
+		"./src/**/*.{js,ts,jsx,tsx}",
+	],
+	theme: {
+		extend: {},
+	},
+	plugins: [],
 };
